requirements.txt
flask
pandas
python-dotenv
google-generativeai
requests
gunicorn
bs4
google-cloud-aiplatform